+++ lib/sloth.dart
@@ -3,3 +3,4 @@
 library sloth;
 
 export 'package:sloth_datetime/sloth_datetime.dart';
+export 'package:sloth_link_crawler/sloth_link_crawler.dart';