+++ packages/sloth_link_crawler/lib/helpers/robots_helper.dart
@@ -1,0 +1,4 @@
+class RobotsHelper {
+  String url;
+  RobotsHelper(this.url);
+}