+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+- ADD version 0.0.2 of sloth_link_crawler
+
 ## 1.0.1+1
 
 - BUMP sloth_datetime version